Import Berkeley DB snapshot, v5.3.21

+++ examples/c/csv/util.c
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2005, 2011 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2005, 2012 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -124,7 +124,7 @@ secondary_callback(DB *db_arg, const DBT *key, const DBT *data, DBT *result)
 	switch (f->type) {
 	case STRING:
 		result->data = *(char **)faddr;
-		result->size = strlen(*(char **)faddr) + 1;
+		result->size = (u_int32_t)strlen(*(char **)faddr) + 1;
 		break;
 	case DOUBLE:
 		if ((addr = malloc(sizeof(double))) == NULL)
