+++ src/common/util_arg.c
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2001, 2011 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2001, 2012 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -49,7 +49,7 @@ __db_util_arg(arg0, str, argcp, argvp)
 		}
 	*ap = NULL;
 
-	*argcp = ap - argv;
+	*argcp = (int)(ap - argv);
 	*argvp = argv;
 
 	return (0);
